requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3